feat(dictation): end a session after a pause on Apple's recognizer too
Hands-free stop was the local engine's alone: watch_for_silence was only
spawned for Engine::Whisper, on the grounds that Apple's analyzer decides
for itself when an utterance ended and we have no PCM to measure. The
second half of that stopped being true — speech_sink already calls
meter.record_buffer on every render-thread buffer, so per-buffer RMS exists
on both paths. What was missing was only the speech-tracking state, which
lived inside Pcm and so belonged to the Whisper sink alone.

So Settings › Dictation › "Stop after a pause" did nothing at all for
anyone on the default engine, while the row sat there reading as on.

Move the tracker to where both taps already meet. Meter::record is the
single funnel — Whisper's tap calls it with the RMS from append_mono,
Apple's record_buffer measures the buffer and calls it — so tracking inside
record covers both paths without new plumbing. The state and the pure
policy (is_speech, rise_factor, settle_floor, track, should_auto_stop) move
verbatim from capture.rs to level.rs as a SpeechTracker owned by Meter; no
algorithm and no constant changed, and the eleven tests move with their
bodies byte-identical.

watch_for_silence now takes the Arc<Meter> that start already holds, which
retires both the on_main hop and captured(). Pcm keeps only rate, limit and
samples: its closed flag was a subset of Meter's, since every path that
cancels a sink also closes the meter.

An unreadable meter never reports a pause. Meter::readable is false for
input layouts that aren't deinterleaved float32, where record_buffer
records nothing; a tracker fed nothing would see "never spoke" and fire
NO_SPEECH_TIMEOUT at ten seconds on a user who is mid-sentence. The gate
lives on Meter rather than the tracker, which has no way to know it is
being starved. Only the Apple path can reach it — capture::sink refuses a
non-float32 format outright.

Also un-splices a doc comment in mod.rs whose head and tail had been
swapped with set_dictation_engine's, and corrects auto_stop()'s: it is read
by remote::status on every platform now, not only by the macOS monitor.
